Final Refactor
Also fixes issues where input is blurred on state change of nested components on Reset page
