+++ .gitignore
@@ -42,3 +42,4 @@ Thumbs.db
 .next
 
 .angular
+.vscode/settings.json
